GH #1 Fixed uninstall and added documentation
Changes include fix for GH #1 and numberous documentation changes.

+++ launchutil.py
@@ -1,6 +1,19 @@
 #!/usr/bin/env python3
 # 
 # Copyright 2023 Charles Y. Choi
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
 import os
 import sys
@@ -288,10 +301,11 @@ class LaunchdUninstall(LaunchdBase):
                                                              service)
 
         if not os.path.exists(target):
-            self.stderr.write('WARNING: "{0}" does not exist.\n'.format(target))
+            self.stderr.write('ERROR: "{0}" does not exist.\n'.format(target))
 
         if self.args.execute:
-            os.unlink(target)
+            if os.path.exists(target):
+                os.unlink(target)
         else:
             self.stderr.write('command: rm {0}\n'.format(target))
             self.stderr.write('add -x or --execute flag to execute command.\n')
